refactor: split TodoList.vue into smaller components
