+++ .eslintrc.json
@@ -2,6 +2,7 @@
   "env": {
     "node": true
   },
+  "parser": "vue-eslint-parser",
   "extends": ["eslint:recommended", "plugin:vue/vue3-recommended", "prettier"],
   "rules": {
     "max-lines": [
