+++ package.json
@@ -8,7 +8,7 @@
     "build": "vite build",
     "preview": "vite preview",
     "format": "npx prettier --write .",
-    "lint:js": "eslint . --ext .js,.vue",
+    "lint:js": "eslint 'src/**/*.{js,vue}'",
     "prepare": "husky install"
   },
   "dependencies": {
@@ -31,7 +31,8 @@
     "prettier": "2.7.1",
     "tailwindcss": "^3.1.8",
     "vfonts": "^0.0.3",
-    "vite": "^3.1.0"
+    "vite": "^3.1.0",
+    "vue-eslint-parser": "^9.1.0"
   },
   "lint-staged": {
     "*.{js,vue}": "eslint --cache --fix",
